+++ game-marvelrivals/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None Planned
 
+## [0.5.1] - 2026-02-03
+
+- Fixed: path strings
+- Improved: Made Root Folder mod installer case-insensitive to folder name
+
 ## [0.5.0] - 2025-12-09
 
 - Fixed issue with Load Order sorting not working if certain other UE game extensions were installed. You will need to reinstall all pak mods to be able to sort them properly. A notification will be sent reminding you to do this.